Don't create pagemarks at page end. Fixes #620.
If a pagemark that ends at the bottom of the page is resized so it
begins below the top, using "create pagemark at point" on a following
page will create a zero-height pagemark on the page and later fail when
the reading progress has multiple pagemarks with percentage 100.

This commit both prevents the creation of such pagemarks and deletes any
existing pagemarks that begin at the end of the page.